requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.0
tensorflow==2.16.1
torch==2.3.0
stable-baselines3==2.3.2
gymnasium==0.29.1
ultralytics==8.2.0
Pillow==10.3.0
pandas==2.2.2
pydantic==2.7.1
gunicorn==22.0.0
